+++ src/client/components/account_emails_list/account_emails_list.js
@@ -7,6 +7,7 @@
  * Parms:
  * - item: a ReactiveVar which holds the account object to edit (may be empty, but not null)
  * - checker: a ReactiveVar which holds the parent Checker
+ * - amInstance: a ReactiveVar which holds the amClass instance
  */
 
 import _ from 'lodash';