+++ src/common/js/functions.js
@@ -4,12 +4,13 @@
 
 /**
  * @param {String} action
+ * @param {amClass} amInstance
  * @param {String} userId
  * @returns {Boolean} true if the current user is allowed to do the action
  */
-AccountsManager.isAllowed = async function( action, userId=null ){
+AccountsManager.isAllowed = async function( action, amInstance, userId=null ){
     let allowed = false;
-    const fn = AccountsManager.configure().allowFn;
+    const fn = amInstance.allowFn();
     if( fn ){
         allowed = await fn( ...arguments );
     }